requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.6  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.57.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis


wagtail==5.2
wagtail-django-recaptcha==1.0
wagtailmenus==3.1.9
wagtail_modeladmin
django-recaptcha==3.0.0
pandas==2.1.3
selenium==4.15.2
XlsxWriter==3.1.9
aiohttp==3.9.1
aiofiles==23.2.1

# don't forget to run npm install
# npm install izitoast --save
# generate model er:
# python manage.py graph_models -a --dot -o myapp_models.dot
# dot -Tpng myapp_models.dot -o myapp_models.png

